requirements.txt
django-extra-views==0.6.4
django-bootstrap-form==3.1